// lunadefend/js/sdks/packages/tokenizer-sdk/src/generated/api.ts
export interface DetokenizeResponse {
    /**
     * 
     * @type {boolean}
     * @memberof DetokenizeResponse
     */
    success: boolean;
    /**
     * 
     * @type {DetokenizeResponseData}
     * @memberof DetokenizeResponse
     */
    data: DetokenizeResponseData;
}
/**
 * 
 * @export
 * @interface DetokenizeResponseData
 */
export interface DetokenizeResponseData {
    /**
     * 
     * @type {string}
     * @memberof DetokenizeResponseData
     */
    downloadUrl: string;
    /**
     * 
     * @type {object}
     * @memberof DetokenizeResponseData
     */
    headers: object;
}
// ...
export interface GetMetadataResponse {
    /**
     * 
     * @type {boolean}
     * @memberof GetMetadataResponse
     */
    success: boolean;
    /**
     * 
     * @type {GetMetadataResponseData}
     * @memberof GetMetadataResponse
     */
    data: GetMetadataResponseData;
}
/**
 * 
 * @export
 * @interface GetMetadataResponseData
 */
export interface GetMetadataResponseData {
    /**
     * 
     * @type {MetaData}
     * @memberof GetMetadataResponseData
     */
    metadata: MetaData;
}
/**
 * @type MetaData
 * The metadata set for the token.  Metadata is immutable once the token is created.  DataType must be set to tell if it is a string or a file that is tokenized.
 * @export
 */
export type MetaData = FileMeta | StringMeta;
// ...
export interface SetGrantResponse {
    /**
     * 
     * @type {boolean}
     * @memberof SetGrantResponse
     */
    success: boolean;
    /**
     * 
     * @type {object}
     * @memberof SetGrantResponse
     */
    data: object;
}
// ...
export interface TokenizeResponse {
    /**
     * 
     * @type {boolean}
     * @memberof TokenizeResponse
     */
    success: boolean;
    /**
     * 
     * @type {TokenizeResponseData}
     * @memberof TokenizeResponse
     */
    data: TokenizeResponseData;
}
/**
 * 
 * @export
 * @interface TokenizeResponseData
 */
export interface TokenizeResponseData {
    /**
     * 
     * @type {string}
     * @memberof TokenizeResponseData
     */
    tokenId: string;
    /**
     * 
     * @type {string}
     * @memberof TokenizeResponseData
     */
    uploadUrl: string;
    /**
     * 
     * @type {object}
     * @memberof TokenizeResponseData
     */
    headers: object;
}